+++ lib/sidekiq/fetch.rb
@@ -2,6 +2,7 @@
 
 require "sidekiq"
 require "sidekiq/component"
+require "sidekiq/capsule"
 
 module Sidekiq # :nodoc:
   class BasicFetch
@@ -26,31 +27,30 @@ module Sidekiq # :nodoc:
       end
     }
 
-    def initialize(config)
-      raise ArgumentError, "missing queue list" unless config[:queues]
-      @config = config
-      @strictly_ordered_queues = !!@config[:strict]
-      @queues = @config[:queues].map { |q| "queue:#{q}" }
+    def initialize(cap)
+      raise ArgumentError, "missing queue list" unless cap.queues
+      @config = cap
+      @strictly_ordered_queues = (config.queues.size == config.queues.uniq.size)
+      @queues = config.queues.map { |q| "queue:#{q}" }
       if @strictly_ordered_queues
         @queues.uniq!
-        @queues << {timeout: TIMEOUT}
       end
     end
 
     def retrieve_work
       qs = queues_cmd
       # 4825 Sidekiq Pro with all queues paused will return an
-      # empty set of queues with a trailing TIMEOUT value.
-      if qs.size <= 1
+      # empty set of queues
+      if qs.size <= 0
         sleep(TIMEOUT)
         return nil
       end
 
-      queue, job = redis { |conn| conn.brpop(*qs) }
+      queue, job = redis { |conn| conn.blocking_call(false, "brpop", *qs, TIMEOUT) }
       UnitOfWork.new(queue, job, config) if queue
     end
 
-    def bulk_requeue(inprogress, options)
+    def bulk_requeue(inprogress)
       return if inprogress.empty?
 
       logger.debug { "Re-queueing terminated jobs" }
@@ -83,7 +83,6 @@ module Sidekiq # :nodoc:
       else
         permute = @queues.shuffle
         permute.uniq!
-        permute << {timeout: TIMEOUT}
         permute
       end
     end