+++ lib/sidekiq/web.rb
@@ -30,12 +30,9 @@ module Sidekiq
       "Queues" => "queues",
       "Retries" => "retries",
       "Scheduled" => "scheduled",
-      "Dead" => "morgue"
+      "Dead" => "morgue",
+      "Metrics" => "metrics"
     }
-
-    if ENV["SIDEKIQ_METRICS_BETA"] == "1"
-      DEFAULT_TABS["Metrics"] = "metrics"
-    end
 
     class << self
       def settings
@@ -77,14 +74,6 @@ module Sidekiq
 
       def set(attribute, value)
         send(:"#{attribute}=", value)
-      end
-
-      def sessions=(val)
-        puts "WARNING: Sidekiq::Web.sessions= is no longer relevant and will be removed in Sidekiq 7.0. #{caller(1..1).first}"
-      end
-
-      def session_secret=(val)
-        puts "WARNING: Sidekiq::Web.session_secret= is no longer relevant and will be removed in Sidekiq 7.0. #{caller(1..1).first}"
       end
 
       attr_accessor :app_url, :redis_pool
@@ -131,10 +120,6 @@ module Sidekiq
 
     def set(attribute, value)
       send(:"#{attribute}=", value)
-    end
-
-    def sessions=(val)
-      puts "Sidekiq::Web#sessions= is no longer relevant and will be removed in Sidekiq 7.0. #{caller[2..2].first}"
     end
 
     def self.register(extension)