+++ myapp/config/application.rb
@@ -7,10 +7,8 @@ require "rails"
   action_view/railtie
   action_mailer/railtie
   active_job/railtie
-  sprockets/railtie
 ].each do |railtie|
   require railtie
-rescue LoadError
 end
 
 # Require the gems listed in Gemfile, including any gems
@@ -23,10 +21,13 @@ module Myapp
     # ensure we work with all defaults.
     config.load_defaults "#{Rails::VERSION::MAJOR}.#{Rails::VERSION::MINOR}"
 
-    # Settings in config/environments/* take precedence over those specified here.
-    # Application configuration can go into files in config/initializers
-    # -- all .rb files in that directory are automatically loaded after loading
-    # the framework and any gems in your application.
+    # Configuration for the application, engines, and railties goes here.
+    #
+    # These settings can be overridden in specific environments using the files
+    # in config/environments, which are processed later.
+    #
+    # config.time_zone = "Central Time (US & Canada)"
+    # config.eager_load_paths << Rails.root.join("extras")
     config.active_job.queue_adapter = :sidekiq
   end
 end