+++ myapp/config/initializers/sidekiq.rb
@@ -16,19 +16,15 @@ Sidekiq.configure_server do |config|
   end
 end
 
-if ENV["SIDEKIQ_REDIS_CLIENT"]
-  Sidekiq::RedisConnection.adapter = :redis_client
-end
-
-class EmptyWorker
-  include Sidekiq::Worker
+class EmptyJob
+  include Sidekiq::Job
 
   def perform
   end
 end
 
-class TimedWorker
-  include Sidekiq::Worker
+class TimedJob
+  include Sidekiq::Job
 
   def perform(start)
     now = Time.now.to_f
@@ -36,27 +32,34 @@ class TimedWorker
   end
 end
 
-Sidekiq::Extensions.enable_delay!
-
-module Myapp
-  class Current < ActiveSupport::CurrentAttributes
-    attribute :tenant_id
-  end
-end
-
 require "sidekiq/middleware/current_attributes"
-Sidekiq::CurrentAttributes.persist(Myapp::Current) # Your AS::CurrentAttributes singleton
+Sidekiq::CurrentAttributes.persist("Myapp::Current") # Your AS::CurrentAttributes singleton
 
 # Sidekiq.transactional_push!
 
 # create a label based on the shorthash and subject line of the latest commit in git.
 # WARNING: you only want to run this ONCE! If this runs on boot for 20 different Sidekiq processes,
-# you will get 20 different deploy marks in Redis! Instead this should go into the script
+# you can get multiple deploy marks in Redis! Instead this should go into the script
 # that runs your deploy, e.g. your capistrano script.
 Sidekiq.configure_server do |config|
   label = `git log -1 --format="%h %s"`.strip
-  require "sidekiq/metrics/deploy"
-  Sidekiq::Metrics::Deploy.new.mark(label: label)
+  require "sidekiq/deploy"
+  Sidekiq::Deploy.mark!(label)
+end
+
+class Singler
+  include Sidekiq::ServerMiddleware
+  def call(w, j, q)
+    puts "single-threaded #{w.class.name}!"
+  end
+end
+
+Sidekiq.configure_server do |config|
+  config.capsule("single_threaded") do |cap|
+    cap.concurrency = 1
+    cap.queues = %w[single]
+    cap.server_middleware.add Singler
+  end
 end
 
 # helper jobs for seeding metrics data